Changed 'id' type form Long to Integer

+++ src/main/java/com/gabchak/repositories/OrderRepository.java
@@ -3,5 +3,5 @@ package com.gabchak.repositories;
 import com.gabchak.models.Order;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OrderRepository extends JpaRepository<Order, Long> {
+public interface OrderRepository extends JpaRepository<Order, Integer> {
 }
